refactor: Create unified BackButton component with prefetching
- Create reusable BackButton component that handles:
  - Smart navigation (history.back() if same-origin, fallback URL otherwise)
  - Prefetching for fallback URLs
  - Icon-only display (no text) for consistent UI
- Update climb-view-actions, settings, and auth pages to use BackButton
- Remove unused BackToClimbList component (dead code)
