+++ packages/web/app/auth/login/auth-page-content.tsx
@@ -2,10 +2,11 @@
 
 import React, { useState, useEffect } from 'react';
 import { Layout, Card, Form, Input, Button, Tabs, Typography, Divider, message, Space } from 'antd';
-import { UserOutlined, LockOutlined, MailOutlined, ArrowLeftOutlined } from '@ant-design/icons';
+import { UserOutlined, LockOutlined, MailOutlined } from '@ant-design/icons';
 import { signIn, useSession } from 'next-auth/react';
 import { useRouter, useSearchParams } from 'next/navigation';
 import Logo from '@/app/components/brand/logo';
+import BackButton from '@/app/components/back-button';
 
 const { Content, Header } = Layout;
 const { Title, Text } = Typography;
@@ -232,7 +233,7 @@ export default function AuthPageContent() {
           boxShadow: '0 1px 2px rgba(0,0,0,0.03)',
         }}
       >
-        <Button icon={<ArrowLeftOutlined />} type="text" onClick={() => router.back()} />
+        <BackButton />
         <Logo size="sm" showText={false} />
         <Title level={4} style={{ margin: 0, flex: 1 }}>
           Welcome
